+++ wnnet/__about__.py
@@ -12,6 +12,6 @@ __all__ = [
 
 __title__ = "wnnet"
 __summary__ = "Python project to handle webnucleo reaction networks"
-__version__ = "2.0.4"
+__version__ = "2.1.0"
 __author__ = "Clemson University"
 __copyright__ = "Clemson University, 2022-2024"